moved classes to an apache package
